% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn.R
\name{AddTargetAssay}
\alias{AddTargetAssay}
\title{Add the TF target expression}
\usage{
AddTargetAssay(object, target.assay = "target", df.grn = NULL)
}
\arguments{
\item{object}{A Seurat object used as input}

\item{target.assay}{Name for the new assay. Default: "target"}

\item{df.grn}{A data frame containing the predicted gene regulatory network.}
}
\value{
A Seurat object with a new assay named by target.assay
}
\description{
This function will created a new assay by using the predicted gene
regulatory network where features are the selected TFs. Each value represents
the average expression of all targets calculated by using the function
\code{\link{AddModuleScore}} from Seurat.
}
