% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_analysis.R
\name{AddTrajectory}
\alias{AddTrajectory}
\title{Add trajectory}
\usage{
AddTrajectory(
  object,
  name = "Trajectory",
  trajectory = NULL,
  group.by = NULL,
  reduction = NULL,
  dims = NULL,
  pre.filter.quantile = 0.9,
  post.filter.quantile = 0.9,
  use.all = FALSE,
  dof = 250,
  spar = 1,
  seed = 42
)
}
\arguments{
\item{object}{A Seurat object}

\item{name}{A string indicating the name of the fitted trajectory to be added in meta.data}

\item{trajectory}{The order of cell groups to be used for constraining
the initial supervised fitting procedure}

\item{group.by}{The column name in metadata that contains the cell group
definitions used in \code{trajectory}}

\item{reduction}{Name of dimension reduction used to infer the trajectory}

\item{dims}{How many dimensions to use}

\item{pre.filter.quantile}{Prior to the initial supervised trajectory fitting,
cells whose euclidean distance from the cell-grouping center is
above the provided quantile will be excluded.}

\item{post.filter.quantile}{After initial supervised trajectory fitting, cells
whose euclidean distance from the cell-grouping center is above the provided
quantile will be excluded.}

\item{use.all}{Whether or not to use cells ourside of trajectory groups for
post-fitting procedure}

\item{dof}{The number of degrees of freedom to be used in the spline fit}

\item{spar}{The sparsity to be used in the spline fit.}

\item{seed}{Random seed}
}
\value{
Seurat object with inferred trajectory
}
\description{
This function infers a pseudo-time trajectory based on the supervised input.
It was modified from the package ArchR \code{\link{addTrajectory}}
to use a Seurat object as input.
For more details, check here \url{https://www.archrproject.com/reference/addTrajectory.html}. \cr
Author: Jeffrey Granja
}
\examples{
\dontrun{
obj <- AddTrajectory(
   object = obj,
   trajectory = c(2, 1),
   group.by = "RNA_snn_res.0.2",
   reduction = "dm",
   dims = 1:3,
   use.all = FALSE
)
}
}
