% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CellPropPlot}
\alias{CellPropPlot}
\title{Plot cell proportion}
\usage{
CellPropPlot(object, group.by = NULL, prop.in = NULL, cols = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Name of one metadata column to group the cells}

\item{prop.in}{Name of one metadata column to compute the cell proportion}

\item{cols}{Specific colors for plotting}
}
\value{
A ggplot object
}
\description{
This function will generate a bar plot to visualize the cell proportion across
different samples
}
\examples{
library(Seurat)
data(pbmc_small)
CellPropPlot(
pbmc_small,
group.by = "RNA_snn_res.1",
prop.in = "groups"
)
}
