% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_cells.R
\name{CoembedData}
\alias{CoembedData}
\title{Integrate single-cell gene expression and chromatin accessibility data.}
\usage{
CoembedData(
  obj.rna,
  obj.atac,
  gene.activity,
  reference.assay = "RNA",
  reduction = "cca",
  weight.reduction = NULL,
  verbose = TRUE
)
}
\arguments{
\item{obj.rna}{A Seurat object including gene expression data}

\item{obj.atac}{A Seurat object including chromatin accessibility data}

\item{gene.activity}{A sparse matrix containing gene activity score per cell}

\item{reference.assay}{The reference assay from gene expression data}

\item{reduction}{The reduction name used for function FindTransferAnchors.}

\item{weight.reduction}{Dimensional reduction to use for the weighting anchors}

\item{verbose}{Print progress bars and output}
}
\value{
An integrated Seurat object
}
\description{
Integrate single-cell gene expression and chromatin accessibility data.
}
\examples{
\dontrun{
obj.coembed <- CoembedData(
   obj.rna = obj.rna,
   obj.atac = obj.atac,
   gene.activity = gene.activity,
   weight.reduction = "harmony"
)

}
}
