% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CompareCellProp}
\alias{CompareCellProp}
\title{Compare the cell proportion}
\usage{
CompareCellProp(
  object,
  group.by = NULL,
  prop.in = NULL,
  sample.annotation = NULL,
  cols = NULL,
  comparisons = NULL,
  method.test = "wilcox.test"
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Name of one metadata column to group the cells, usually this refers to
the clustering results.}

\item{prop.in}{Name of one metadata column to compute the cell proportion, usually
this refers to the sample identify}

\item{sample.annotation}{Name of one metadata column to group the samples for comparison}

\item{cols}{Specific colors for plotting}

\item{comparisons}{A list of length-2 vectors used to compare the proportion.
This parameter is passed to the function stat_compare_means.}

\item{method.test}{Name of method for statistical test. Default: "wilcox.test"}
}
\value{
A ggplot object
}
\description{
Compare the cell proportion
}
\examples{
\dontrun{
p <- CompareCellProp(
   object = coembed,
   group.by = "RNA_snn_res.0.2",
   prop.in = "patient_region_id",
   sample.annotation = "patient_group",
    comparisons = list(c("myogenic", "ischemic"),
                       c("ischemic", "fibrotic"),
                       c("myogenic", "fibrotic"))
)
}
}
