% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CorrelationHeatmap}
\alias{CorrelationHeatmap}
\title{Plot correlation heatmap}
\usage{
CorrelationHeatmap(trajectory1, trajectory2, df_cor)
}
\arguments{
\item{trajectory1}{A SummarizedExperiment object for TF activity
generated by the function \code{\link{GetTrajectory}}}

\item{trajectory2}{A SummarizedExperiment object for TF expression
generated by the function \code{\link{GetTrajectory}}}

\item{df_cor}{A data frame containing the correlation between TF expression and
TF activity}
}
\value{
A heatmap
}
\description{
This function will generate a combined heatmap with each one corresponding to
the TF expression and binding activity as selected by using the function
\code{\link{GetCorrelation}}. It was modified from the package \code{ArchR}.
More details can be found:
\url{https://www.archrproject.com/bookdown/myeloid-trajectory-monocyte-differentiation.html}
}
