% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CorrelationPlot}
\alias{CorrelationPlot}
\title{Plot correlation}
\usage{
CorrelationPlot(df)
}
\arguments{
\item{df}{A data frame generated by the function \code{\link{GetCorrelation}}}
}
\value{
A ggplot object
}
\description{
This function will generate a scatter plot to visualize the correlation between
expression and activity for all TFs passing the threshold.
}
