% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_cells.R
\name{CreatePairedObject}
\alias{CreatePairedObject}
\title{Create paired object}
\usage{
CreatePairedObject(
  df.pair,
  obj.coembed = NULL,
  obj.rna = NULL,
  obj.atac = NULL,
  rna.assay = NULL,
  atac.assay = NULL,
  sep = c("-", "-")
)
}
\arguments{
\item{df.pair}{A data frame containing the cell pairing results
generated by the function \code{\link{PairCells}}.}

\item{sep}{Separators to use for strings encoding genomic coordinates.
First element is used to separate the chromosome from the coordinates,
second element is used to separate the start from end coordinate.}

\item{object}{Seurat object generated by the function \code{\link{CoembedData}}}

\item{use.assay1}{A string indicating the first assay}

\item{use.assay2}{A string indicating the first assay}
}
\value{
A Seurat object
}
\description{
This function will create a Seurat object including single-cell multimodal
data based on the paired cells.
}
\examples{
\dontrun{
obj <- CreatePairedObject(
   df.pair = df.pair,
   object = obj.coembed,
   use.assay1 = "RNA",
   use.assay2 = "ATAC",
)
}
}
