% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_analysis.R
\name{GetCorrelation}
\alias{GetCorrelation}
\title{Get correlation}
\usage{
GetCorrelation(trajectory1, trajectory2)
}
\arguments{
\item{trajectory1}{A SummarizedExperiment object for TF activity
generated by the function \code{\link{GetTrajectory}}}

\item{trajectory2}{A SummarizedExperiment object for TF expression
generated by the function \code{\link{GetTrajectory}}}
}
\value{
A data frame cantaining the correlation
}
\description{
This function will compute the correlation between TF expression as measured
by RNA-seq and binding activity estimated from ATAC-seq along the trajectory.
}
\examples{
\dontrun{
df.cor <- GetCorrelation(
   trajectory1 = trajMM,
   trajectory2 = trajRNA
)
}
}
