% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn.R
\name{GetGRN}
\alias{GetGRN}
\title{Get gene regulatory network}
\usage{
GetGRN(motif.matching = NULL, df.cor = NULL, df.p2g = NULL)
}
\arguments{
\item{motif.matching}{A matrix of peak by motif to indicate if a peak is bound
by a motif. This matrix should only contain 0 and 1 with 1 indicating a binding
event and 0 indicating no binding site(s).}

\item{df.cor}{A matrix of TF-gene correlation as generated by using the function
\code{\link{GetTFGeneCorrelation}}.}

\item{df.p2g}{A data frame containing predicted peak-to-gene links as generated
by using the function \code{\link{PeakToGene}}.}
}
\value{
A data frame representing gene regulatory network
}
\description{
This function will generate the final prediction of TF-gene network. It takes the
TF-gene correlation. peak-TF binding prediction, and peak-to-gene links as input.
}
