% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_cells.R
\name{GetPairList}
\alias{GetPairList}
\title{Get list of paired cells}
\usage{
GetPairList(cell_matches, ATAC_barcodes, RNA_barcodes)
}
\arguments{
\item{cell_matches}{The output object of \code{\link{fullmatch}} in the package optmatch}

\item{ATAC_barcodes}{Barcode of ATAC cells, must match the order of IDs in cell_matches.}

\item{RNA_barcodes}{Barcode of RNA cells, must match the order of IDs in cell_matches.}
}
\value{
A data frame containing the cell pairs
}
\description{
This function takes in the output of the \code{\link{fullmatch}} function and sort the
results into a list of ATAC-RNA pairs.
It was modified from \url{https://github.com/buenrostrolab/stimATAC_analyses_code/blob/master/R/optMatching_functions.R}
}
