% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_analysis.R
\name{GetTrajectory}
\alias{GetTrajectory}
\title{Get data along the trajectory}
\usage{
GetTrajectory(
  object = NULL,
  trajectory.name = "Trajectory",
  assay = NULL,
  slot = "counts",
  groupEvery = 1,
  log2Norm = TRUE,
  scaleTo = 10000,
  smoothWindow = 11
)
}
\arguments{
\item{object}{Seurat object}

\item{trajectory.name}{The name of trajectory inferred by \code{\link{AddTrajectory}}}

\item{assay}{Which assay is used for data collection}

\item{slot}{Name of slot used to collect data. Usually set as "counts" or "data".}

\item{groupEvery}{The number of sequential percentiles to group together when generating a trajectory.
This is similar to smoothing via a non-overlapping sliding window across pseudo-time.}

\item{log2Norm}{Whether or not the data should be log2 transformed. This can be set to TRUE
if gene expression or chromatin accessibility data is used,
otherwise FALSE if data from TF activity is used.}

\item{scaleTo}{Once the sequential trajectory matrix is created,
each column in that matrix will be normalized to a column sum indicated by scaleTo.}

\item{smoothWindow}{An integer value indicating the smoothing window in size
(relative to groupEvery) for the sequential trajectory matrix to better reveal
temporal dynamics.}
}
\value{
A SummarizedExperiment object
}
\description{
This function collects and smooths data along the trajectory using Seurat
object as input.
It was modified from the package ArchR \code{\link{getTrajectory}}
to use a Seurat object as input.
For more details, check here \url{https://www.archrproject.com/reference/getTrajectory.html}.
}
\examples{
\dontrun{
trajRNA <- GetTrajectory(
   object = obj,
   assay = "RNA",
   slot = "data",
   smoothWindow = 7,
   log2Norm = TRUE
)
}
}
