% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn.R
\name{GRNHeatmap}
\alias{GRNHeatmap}
\title{Get a heatmap of TF gene correlation}
\usage{
GRNHeatmap(df_cor)
}
\arguments{
\item{tf.gene.cor}{A matrix representing TF-gene correlation}

\item{tf.timepoint}{A list of TF time point along the trajectory}

\item{km}{Number of clusters}
}
\value{
A heatmap
}
\description{
This function will generate a heatmap to visualize the TF-gene correlation computed
by the \code{\link{GetTFGeneCorrelation}}
}
