% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn.R
\name{GRNSpatialPlot}
\alias{GRNSpatialPlot}
\title{Plot the target gene in spatial space}
\usage{
GRNSpatialPlot(
  object,
  assay,
  df.grn,
  tf.use,
  min.cutoff = "q5",
  max.cutoff = "q95",
  vis.option = "B",
  crop = TRUE
)
}
\arguments{
\item{object}{A Seurat object of spatial transcriptome data}

\item{assay}{Which assay to plot}

\item{df.grn}{A data frame including the inferred gene regulatory network}

\item{tf.use}{Which TF to use}

\item{min.cutoff}{Minimum cutoff values for each feature}

\item{max.cutoff}{Maximum cutoff values for each feature}

\item{vis.option}{Options for visualization. Default: "B"}
}
\value{
A ggplot object
}
\description{
This function plots the overall expression of all target genes
for a particular TF using spatial transcriptome data. It is based on
}
