% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn.R
\name{NetCentPlot}
\alias{NetCentPlot}
\title{Centric Plot}
\usage{
NetCentPlot(netobj, gene, highlights = NULL)
}
\arguments{
\item{netobj}{scMEGA net}
}
\value{
a ggraph plot
}
\description{
This function will generate a plot centered in a specific gene/TF
}
