% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_cells.R
\name{PairCells}
\alias{PairCells}
\title{Pair the cells}
\usage{
PairCells(
  object,
  reduction = NULL,
  pair.by = NULL,
  ident1 = "ATAC",
  ident2 = "RNA",
  assay = "RNA",
  pair.mode = "geodesic",
  tol = 1e-04,
  search.range = 0.2,
  max.multimatch = 5,
  min_subgraph_size = 50,
  seed = 42,
  k = 300
)
}
\arguments{
\item{object}{The integrated Seurat object generated by the function CoembedData}

\item{reduction}{Dimensional reduction to use for the pairing cells}

\item{pair.by}{Name of one metadata column to split the object for pairing;}

\item{ident1}{Specify how to split the object, must be an value of pair.by}

\item{ident2}{Specify how to split the object, must be an value of pair.by}

\item{assay}{Assay name based on which a KNN graph is constructed}

\item{pair.mode}{Pair mode. Currently only "geodesic" is available}

\item{tol}{Tol times the number of subjects to be matched specifies the extent to which fullmatch's}

\item{search.range}{This determines the size of the search knn. search_range * total number of cells = size of knn.}

\item{max.multimatch}{Maximum number of cells allowed to be matched to each cell}

\item{min_subgraph_size}{Minimum number of cells (ATAC/RNA each) needed for
pairing in a given subgraph. Will skip subgraphs with fewer than these cells}

\item{seed}{Random seed}

\item{k}{k-NN parameter used for applying constraints on ATAC-RNA pairs}
}
\value{
A data frame containing the cell pairs
}
\description{
This function will create the pairs of cells between different modalities based on
the integrated data generated by \code{\link{CoembedData}}.
It was modified from \url{https://github.com/buenrostrolab/stimATAC_analyses_code/blob/master/R/optMatching_functions.R}
to use Seurat object as input.
}
\examples{
\dontrun{
df.pair <- PairCells(
  object = obj.coembed,
  reduction = "harmony",
  pair.by = "tech",
  ident1 = "ATAC",
  ident2 = "RNA"
  )
}
}
