% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_to_gene.R
\name{PeakToGene}
\alias{PeakToGene}
\title{Get peak to gene links}
\usage{
PeakToGene(peak.mat, gene.mat, genome = "hg19", max.dist = 250000)
}
\arguments{
\item{peak.mat}{A matrix containing chromatin accessibility data}

\item{gene.mat}{A matrix containing gene expression data}

\item{genome}{Which genome to use. Currently available are: hg19, hg38, mm9,
and mm10}

\item{max.dist}{The maximum distance between a peak and a gene}
}
\value{
A data frame containing all peak-to-gene links
}
\description{
This function will link peak to genes based on correlation analysis.
}
