% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{RunDiffusionMap}
\alias{RunDiffusionMap}
\title{Run diffusion map}
\usage{
RunDiffusionMap(
  object,
  reduction.key = "DC_",
  assay = DefaultAssay(object = object),
  reduction = NULL,
  dims = NULL,
  k = NULL,
  seed = 42,
  verbose = TRUE,
  reduction.name = "dm"
)
}
\arguments{
\item{object}{A Seurat object}

\item{reduction.key}{dimensional reduction key,
specifies the string before the number for the dimension names. DC by default}

\item{assay}{Assay to pull data}

\item{reduction}{Which dimensional reduction to use for the diffusion map input}

\item{dims}{Which dimensions to use as input features}

\item{k}{Number of nearest neighbors to consider. This is passed to the function
DiffusionMap}

\item{seed}{Random seed}

\item{verbose}{Whether or not print running message}

\item{reduction.name}{Name to store dimensional reduction under in the Seurat object.
Default: dm}
}
\value{
Returns a Seurat object containing a diffusion map representation
}
\description{
This function is a wrapper to run the diffusion map dimensional reduction
technique via package destiny. Details on this package can be found here:
\url{https://bioconductor.org/packages/release/bioc/html/destiny.html}.
}
\examples{
\dontrun{
data("pbmc_small")
# Run diffusion map on first 5 PCs
pbmc_small <- RunDiffusionMap(object = pbmc_small, reduction = "pca",
dims = 1:20, k = 10)
# Plot results
DimPlot(object = pbmc_small, reduction = 'dm')
}
}
