% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_tf_gene.R
\name{SelectGenes}
\alias{SelectGenes}
\title{Select genes}
\usage{
SelectGenes(
  object,
  atac.assay = "ATAC",
  rna.assay = "RNA",
  var.cutoff.gene = 0.9,
  trajectory.name = "Trajectory",
  distance.cutoff = 2000,
  cor.cutoff = 0,
  fdr.cutoff = 1e-04,
  return.heatmap = TRUE,
  labelTop1 = 10,
  labelTop2 = 10
)
}
\arguments{
\item{object}{A Seurat object}

\item{atac.assay}{The assay name for chromatin accessibility. Default: "ATAC"}

\item{rna.assay}{The assay name for gene expression. Default: "RNA"}

\item{var.cutoff.gene}{The cutoff of variation to select genes. Default: 0.9}

\item{trajectory.name}{The trajectory name used for computing correlation
between TF binding activity and TF expression}

\item{distance.cutoff}{The minimum distance between cis-regulatory elements and genes}

\item{cor.cutoff}{The cutoff of peak-to-gene correlation. Default: 0}

\item{fdr.cutoff}{The cutoff of peak-to-gene p-value Default: 1e-04}

\item{return.heatmap}{Whether or not return the heatmap for visualization}

\item{labelTop1}{Number of labels for row names}

\item{labelTop2}{Number of labels for row names}
}
\value{
A list containing a dataframe of peak-to-gene links and a heatmap
}
\description{
Select genes
}
