% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_tf_gene.R
\name{SelectTFs}
\alias{SelectTFs}
\title{Select TFs for GRN inference}
\usage{
SelectTFs(
  object,
  tf.assay = "chromvar",
  rna.assay = "RNA",
  atac.assay = "ATAC",
  trajectory.name = "Trajectory",
  groupEvery = 1,
  p.cutoff = 0.01,
  cor.cutoff = 0.3,
  return.heatmap = TRUE
)
}
\arguments{
\item{object}{A Seurat object}

\item{tf.assay}{The assay name for TF activity. Default: "chromvar"}

\item{rna.assay}{The assay name for gene expression. Default: "RNA"}

\item{atac.assay}{The assay name for Peaks. Default: "ATAC"}

\item{trajectory.name}{The trajectory name used for computing correlation
between TF binding activity and TF expression}

\item{groupEvery}{The number of sequential percentiles to group together when generating a trajectory.
This is similar to smoothing via a non-overlapping sliding window across pseudo-time.}

\item{p.cutoff}{A cutoff of p-values. Default: 0.01}

\item{cor.cutoff}{A cutoff of correlation. Default: 0.3}

\item{return.heatmap}{Whether or not return the heatmap for visualization}
}
\value{
A list containing a dataframe of TF activity and expression correlation
and a heatmap
}
\description{
Select TFs for GRN inference
}
