% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn.R
\name{TopEmbGRN}
\alias{TopEmbGRN}
\title{PCA of Topological Measures}
\usage{
TopEmbGRN(df.grn, gene.cluster = NULL, axis = c(1, 2))
}
\arguments{
\item{df.grn}{A data frame representing predicted network}

\item{genes.cluster}{A data frame containing clustering results of genes}
}
\value{
a prcomp output
}
\description{
This function will generate a graph centrality PCA embedding
}
