% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{TrajectoryHeatmap}
\alias{TrajectoryHeatmap}
\title{Plot trajectory heatmap}
\usage{
TrajectoryHeatmap(
  trajectory,
  varCutOff = 0.9,
  maxFeatures = 25000,
  scaleRows = TRUE,
  rowOrder = NULL,
  limits = c(-1.5, 1.5),
  labelRows = FALSE,
  pal = NULL,
  labelMarkers = NULL,
  labelTop = 50,
  name = "Heatmap",
  returnMatrix = FALSE
)
}
\arguments{
\item{trajectory}{A SummarizedExperiment object that results from calling \code{\link{GetTrajectory}}}

\item{varCutOff}{The "Variance Quantile Cutoff" to be used for identifying
the top variable features across the given trajectory.
Only features with a variance above the provided quantile will be retained.}

\item{maxFeatures}{The maximum number of features, ordered by variance,
to consider from useMatrix when generating a trajectory}

\item{scaleRows}{A boolean value that indicates whether row-wise z-scores
should be computed on the matrix}

\item{rowOrder}{If wanting to set the order of rows to be plotted, the
indices (integer or character correpsonding to rownmaes) can be provided here.}

\item{limits}{A numeric vector of two numbers that represent the lower and
upper limits of the heatmap color scheme.}

\item{labelRows}{A boolean value that indicates whether all rows should be
labeled on the side of the heatmap.}

\item{pal}{A custom continuous palette used to override the default
continuous palette for the heatmap.}

\item{labelMarkers}{A character vector listing the rownames that should be
labeled on the side of the heatmap.}

\item{labelTop}{A number indicating how many of the top N features,
based on variance of the matrix}

\item{name}{Name of the matrix}

\item{returnMatrix}{A boolean value that indicates whether the final heatmap
matrix should be returned in lieu of plotting the actual heatmap.}
}
\value{
a heatmap
}
\description{
This function generates a heatmap to visualize the data along inferred trajectory.
It was modified from the package ArchR \code{\link{plotTrajectoryHeatmap}}.
For more details, check here \url{https://www.archrproject.com/reference/plotTrajectoryHeatmap.html}.
}
