% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{TrajectoryPlot}
\alias{TrajectoryPlot}
\title{Plot trajectory}
\usage{
TrajectoryPlot(
  object = NULL,
  trajectory = "Trajectory",
  reduction = NULL,
  size = 0.2,
  rastr = FALSE,
  quantCut = c(0.01, 0.99),
  continuousSet = NULL,
  discreteSet = NULL,
  randomize = TRUE,
  keepAxis = FALSE,
  baseSize = 6,
  addArrow = FALSE,
  smoothWindow = 5
)
}
\arguments{
\item{object}{Seurat object}

\item{trajectory}{Name of one metadata column to group the cells, usually this refers to
the clustering results}

\item{reduction}{Which dimension reduction is used to visualize the trajectory}

\item{size}{A number indicating the size of the points to plot}

\item{rastr}{Whether the plot should be rasterized}

\item{quantCut}{If this is not NULL, a quantile cut is performed to threshold
the top and bottom of the distribution of numerical values.
This prevents skewed color scales caused by strong outliers.
The format of this should be c(x,y) where x is the lower threshold and y is
the upper threshold. For example, quantileCut = c(0.025,0.975) will take the
2.5th percentile and 97.5 percentile of values and set values below/above to
the value of the 2.5th and 97.5th percentile values respectively.}

\item{continuousSet}{The name of a continuous palette from ArchRPalettes
for visualizing colorBy in the embedding if a continuous color set is desired.}

\item{discreteSet}{The name of a discrete palette from ArchRPalettes for
visualizing colorBy in the embedding if a discrete color set is desired.}

\item{randomize}{A boolean value that indicates whether to randomize points
prior to plotting to prevent cells from one cluster being present at the front of the plot.}

\item{keepAxis}{A boolean value that indicates whether the x and y
axis ticks and labels should be plotted.}

\item{baseSize}{The base font size to use in the plot.}

\item{addArrow}{A boolean value that indicates whether to add a smoothed
arrow in the embedding based on the aligned trajectory.}

\item{smoothWindow}{An integer value indicating the
smoothing window for creating inferred Arrow overlay on to embedding.}
}
\value{
A ggplot object
}
\description{
This function will generate a scatter plot to visualize the inferred trajectory.
It was modified from the package ArchR \code{\link{plotTrajectory}}
to use a Seurat object as input.
For more details, check here \url{https://www.archrproject.com/reference/plotTrajectory.html}.
}
\examples{
\dontrun{
p <- TrajectoryPlot(
   object = obj,
   reduction = "dm",
   continuousSet = "blueYellow",
   size = 1
)
}
}
