% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_cells.R
\name{CoembedData}
\alias{CoembedData}
\title{Integrate data}
\usage{
CoembedData(
  obj.rna,
  obj.atac,
  gene.activity,
  reference.assay = "RNA",
  reduction = "cca",
  weight.reduction = NULL,
  verbose = TRUE,
  dims = 1:30
)
}
\arguments{
\item{obj.rna}{A Seurat object including gene expression data}

\item{obj.atac}{A Seurat object including chromatin accessibility data}

\item{gene.activity}{A sparse matrix containing gene activity score per cell}

\item{reference.assay}{The reference assay from gene expression data}

\item{reduction}{The reduction name used for function FindTransferAnchors.}

\item{weight.reduction}{Dimensional reduction to use for the weighting anchors}

\item{verbose}{Print progress bars and output}

\item{dims}{Set of dimensions to use in the anchor weighting procedure.}
}
\value{
An integrated Seurat object
}
\description{
This function wraps a number of functions from the Seurat package.
It integrates and projects the single-cell multi-omics (e.g., scRNA-seq
and scATAC-seq) data into a common low-dimensional space via CCA approach.
The details can be found:
\url{https://satijalab.org/seurat/articles/atacseq_integration_vignette.html}
}
\examples{
\dontrun{
obj.coembed <- CoembedData(
   obj.rna = obj.rna,
   obj.atac = obj.atac,
   gene.activity = gene.activity,
   weight.reduction = "harmony"
)

}
}
