% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CorrelationHeatmap}
\alias{CorrelationHeatmap}
\title{Plot correlation heatmap}
\usage{
CorrelationHeatmap(
  trajectory1,
  trajectory2,
  name1 = NULL,
  name2 = NULL,
  labelRows1 = TRUE,
  labelRows2 = TRUE,
  labelTop1 = 50,
  labelTop2 = 50,
  limits1 = c(-2, 2),
  limits2 = c(-2, 2)
)
}
\arguments{
\item{trajectory1}{A SummarizedExperiment object for TF activity
generated by the function \code{\link{GetTrajectory}}}

\item{trajectory2}{A SummarizedExperiment object for TF expression}

\item{name1}{Name for the first object}

\item{name2}{Name for the second object}

\item{labelRows1}{Whether or not to show all labels for the first heatmap}

\item{labelRows2}{Whether or not to show all labels for the second heatmap}

\item{labelTop1}{A number indicating how many of the top N features,
based on variance, should be labeled on the side of the first heatmap}

\item{labelTop2}{A number indicating how many of the top N features,
based on variance, should be labeled on the side of the second heatmap}

\item{limits1}{A numeric vector of two numbers that represent the lower and
upper limits of the second heatmap color scheme.
generated by the function \code{\link{GetTrajectory}}}

\item{limits2}{A numeric vector of two numbers that represent the lower and
upper limits of the second heatmap color scheme.
generated by the function \code{\link{GetTrajectory}}}
}
\value{
A heatmap
}
\description{
This function will generate a combined heatmap with each one corresponding to
the TF expression and binding activity as selected by using the function
\code{\link{GetCorrelation}}. It was modified from the package \code{ArchR}.
More details can be found:
\url{https://www.archrproject.com/bookdown/myeloid-trajectory-monocyte-differentiation.html}
}
