% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grn.R
\name{GetTFGeneCorrelation}
\alias{GetTFGeneCorrelation}
\title{Get TF gene correlation}
\usage{
GetTFGeneCorrelation(
  object,
  tf.use = NULL,
  gene.use = NULL,
  tf.assay = "chromvar",
  gene.assay = "RNA",
  atac.assay = "ATAC",
  trajectory.name = "Trajectory",
  groupEvery = 1
)
}
\arguments{
\item{object}{A Seurat object}

\item{tf.use}{A string list to specify which TFs to use for correlation computation}

\item{gene.use}{A string list to specify which genes to use for correlation computation}

\item{tf.assay}{Assay that includes TF activity data. Default: "chromvar"}

\item{gene.assay}{Assay that includes gene expression activity data. Default: "RNA"}

\item{atac.assay}{Assay that includes peaks data. Default: "ATAC"}

\item{trajectory.name}{Trajectory name}

\item{groupEvery}{The number of sequential percentiles to group together when generating a trajectory.
This is similar to smoothing via a non-overlapping sliding window across pseudo-time.}
}
\value{
A matrix containing TF-gene correlation
}
\description{
This function will compute the correlation between TF binding activity and
gene expression along the trajectory.
}
