% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_cells.R
\name{PairCells}
\alias{PairCells}
\title{Pair the cells between different modalities based on the coembedding space}
\usage{
PairCells(
  object,
  reduction = NULL,
  pair.by = NULL,
  ident1 = "ATAC",
  ident2 = "RNA",
  assay = "RNA",
  pair.mode = "geodesic",
  tol = 1e-04,
  search.range = 0.2,
  max.multimatch = 5,
  min_subgraph_size = 50,
  seed = 42,
  k = 300
)
}
\arguments{
\item{object}{The integrated Seurat object generated by the function CoembedData}

\item{reduction}{Dimensional reduction to use for the pairing cells}

\item{pair.by}{Name of one metadata column to split the object for pairing;}

\item{ident1}{Specify how to split the object, must be an value of pair.by}

\item{ident2}{Specify how to split the object, must be an value of pair.by}

\item{assay}{Assay name based on which a KNN graph is constructed}

\item{pair.mode}{Pair mode. Currently only "geodesic" is available}

\item{k}{}
}
\value{
A data frame containing the cell pairs
}
\description{
Pair the cells between different modalities based on the coembedding space
}
\examples{
\dontrun{
df.pair <- PairCells(
  object = obj.coembed,
  reduction = "harmony",
  pair.by = "tech",
  ident1 = "ATAC",
  ident2 = "RNA"
  )
}
}
