% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PseudotimePlot}
\alias{PseudotimePlot}
\title{Plot dynamic changes along pseudotime trajectory}
\usage{
PseudotimePlot(
  object,
  tf.use,
  tf.assay = "chromvar",
  rna.assay = "RNA",
  atac.assay = "ATAC",
  target.assay = "target",
  trajectory.name = "Trajectory"
)
}
\arguments{
\item{object}{A Seurat object used as input.}

\item{tf.use}{Which TF to plot.}

\item{tf.assay}{Assay name for TF binding activity. Default: "chromvar"}

\item{rna.assay}{Assay name for TF expression. Default: "RNA"}

\item{atac.assay}{Assay name for Peaks. Default: "ATAC"}

\item{target.assay}{Assay name for TF target expression. Default: "target"}

\item{trajectory.name}{Trajectory name for visualization.}
}
\value{
A ggplot object
}
\description{
This function will generate a line plot to visualize the TF binding activity,
TF expression, and TF target expression along the trajectory.
}
